chore: separate props from HvRoot and HvScreen (#791)
In current release, HvRoot is sharing the same props as HvScreen. Some
recent required props for HvScreen are causing HvRoot to include
required props which it doesn't need.

Separating the two props out allows HvRoot to only require the external
dependencies.

Asana: https://app.asana.com/0/1204008699308084/1206433111684035/f
